requirements.txt
openai
flask
python-dotenv
gunicorn